test: consolidate ad30 local ipc depth coverage
