+++ crates/atm-daemon/src/local_ipc_transport.rs
@@ -57,10 +57,6 @@ pub(crate) const CONNECTION_WORKER_PANIC_RECOVERED_MESSAGE: &str =
     "daemon local IPC connection worker panicked; transport thread recovered";
 pub(crate) const DISPATCH_PANIC_RECOVERED_MESSAGE: &str =
     "daemon local IPC dispatch worker panicked before completing; transport thread recovered";
-// Test hooks keep shutdown deadlines short so the transport suite verifies
-// drain/cancel behavior without waiting on production-scale timing.
-#[cfg(all(test, unix))]
-const TEST_GRACEFUL_DRAIN_DEADLINE: Duration = Duration::from_millis(500);
 const TERMINATE_REJECTION_REQUEST_ID: u64 = NonZeroU64::MIN.get();
 
 #[derive(Debug, Default)]
@@ -282,16 +278,6 @@ impl std::fmt::Debug for PreparedRuntimeServer {
 }
 
 impl PreparedRuntimeServer {
-    #[cfg(all(test, unix))]
-    fn bind(endpoint_path: PathBuf) -> Result<Self, AtmError> {
-        Self::bind_with_observability(
-            endpoint_path,
-            SubsystemObservability::disabled(DaemonSubsystem::LocalIpcTransport),
-            SubsystemObservability::disabled(DaemonSubsystem::HostOwnership),
-            SubsystemObservability::disabled(DaemonSubsystem::LifecycleControl),
-        )
-    }
-
     fn bind_with_observability(
         endpoint_path: PathBuf,
         observability: SubsystemObservability,
@@ -850,25 +836,6 @@ impl boundary::ServerTransport for LocalIpcServerTransportAdapter {
 mod tests {
     use super::*;
     use crate::active_connection_registry::TrackedDispatchHandle;
-    #[cfg(unix)]
-    use crate::lifecycle_control::LifecycleControlSourceAdapter;
-    #[cfg(unix)]
-    use crate::test_support::{
-        DoctorOnlyDispatcher, LifecycleFlagResetGuard, PanicDispatcher,
-        connect_daemon_local_ipc_until_ready,
-    };
-    #[cfg(unix)]
-    use atm_core::boundary::RequestDispatcher;
-    #[cfg(unix)]
-    use atm_core::doctor::DoctorQuery;
-    #[cfg(unix)]
-    use atm_core::error_codes::AtmErrorCode;
-    #[cfg(unix)]
-    use atm_core::protocol::{RequestEnvelope, ResponseEnvelope};
-    #[cfg(unix)]
-    use atm_core::test_support::EnvGuard;
-    #[cfg(unix)]
-    use std::sync::Arc;
     use std::sync::atomic::Ordering;
     use std::sync::mpsc;
     use std::time::{Duration, Instant};
@@ -920,260 +887,6 @@ mod tests {
             error
                 .to_string()
                 .contains("failed to create daemon local IPC directory")
-        );
-    }
-
-    #[cfg(unix)]
-    #[test]
-    #[serial_test::serial(env)]
-    fn accept_error_without_lifecycle_signal_exits_within_one_second() {
-        let tempdir = TempDir::new().expect("tempdir");
-        let atm_home = tempdir.path().join("atm-home");
-        std::fs::create_dir_all(&atm_home).expect("atm home dir");
-        let _env = EnvGuard::set_many([
-            ("ATM_HOME", Some(atm_home.to_str().expect("utf8 atm home"))),
-            ("ATM_LOG_DIR", None),
-            ("HOME", Some(tempdir.path().to_str().expect("utf8 home"))),
-            ("USERPROFILE", None),
-        ]);
-        let socket_path = tempdir.path().join("daemon.sock");
-        let mut runtime = PreparedRuntimeServer::bind(socket_path).expect("prepare runtime");
-        let endpoint_guard = runtime.take_endpoint_guard().expect("take endpoint guard");
-        let lifecycle = LifecycleControlSourceAdapter::install().expect("install lifecycle");
-        let _reset = LifecycleFlagResetGuard::install(lifecycle);
-        let dispatcher: Arc<dyn RequestDispatcher + Send + Sync> = Arc::new(DoctorOnlyDispatcher);
-        let (serve_result_tx, serve_result_rx) = mpsc::channel();
-        let (inject_tx, inject_rx) = mpsc::sync_channel(1);
-        install_injected_accept_error_for_test(&mut runtime, inject_tx);
-
-        let join = std::thread::spawn(move || {
-            // These shortened deadlines validate immediate shutdown-beacon wake correctness, not
-            // the production SLO values documented for the real daemon runtime.
-            let result = runtime.serve_with_runtime_hooks(
-                dispatcher,
-                RuntimeServeHooks {
-                    endpoint_guard,
-                    graceful_drain_deadline: TEST_GRACEFUL_DRAIN_DEADLINE,
-                    force_cancel_deadline: Duration::from_secs(2),
-                    begin_shutdown: || Ok(()),
-                    reload_runtime_view: || Ok(()),
-                    finalize_shutdown: || {},
-                    publish_ready: || Ok(()),
-                },
-            );
-            serve_result_tx.send(result).expect("send serve result");
-        });
-
-        inject_rx
-            .recv_timeout(Duration::from_secs(1))
-            .expect("accept error should inject within 1s");
-        let shutdown_started = Instant::now();
-        let error = serve_result_rx
-            .recv_timeout(Duration::from_secs(1))
-            .expect("serve result should arrive within 1s")
-            .expect_err("serve should fail after injected accept error");
-        assert!(
-            shutdown_started.elapsed() <= Duration::from_secs(1),
-            "lifecycle waiter should observe the shutdown beacon and exit within 1s"
-        );
-        assert!(error.message.contains("accept error") || error.message.contains("accepting"));
-        join.join().expect("join serve thread");
-    }
-
-    #[cfg(unix)]
-    #[test]
-    #[serial_test::serial(env)]
-    fn accept_after_terminate_returns_typed_shutdown_error() {
-        let tempdir = TempDir::new().expect("tempdir");
-        let atm_home = tempdir.path().join("atm-home");
-        std::fs::create_dir_all(&atm_home).expect("atm home dir");
-        let _env = EnvGuard::set_many([
-            ("ATM_HOME", Some(atm_home.to_str().expect("utf8 atm home"))),
-            ("ATM_LOG_DIR", None),
-            ("HOME", Some(tempdir.path().to_str().expect("utf8 home"))),
-            ("USERPROFILE", None),
-        ]);
-        let socket_path = tempdir.path().join("daemon.sock");
-        let mut runtime =
-            PreparedRuntimeServer::bind(socket_path.clone()).expect("prepare runtime");
-        let endpoint_guard = runtime.take_endpoint_guard().expect("take endpoint guard");
-        let lifecycle = LifecycleControlSourceAdapter::install().expect("install lifecycle");
-        let _reset = LifecycleFlagResetGuard::install(lifecycle.clone());
-        let dispatcher: Arc<dyn RequestDispatcher + Send + Sync> = Arc::new(DoctorOnlyDispatcher);
-        let (serve_result_tx, serve_result_rx) = mpsc::channel();
-        let (ready_tx, ready_rx) = mpsc::sync_channel(1);
-
-        let join = std::thread::spawn(move || {
-            // These shortened deadlines validate immediate shutdown-beacon wake correctness, not
-            // the production SLO values documented for the real daemon runtime.
-            let result = runtime.serve_with_runtime_hooks(
-                dispatcher,
-                RuntimeServeHooks {
-                    endpoint_guard,
-                    graceful_drain_deadline: TEST_GRACEFUL_DRAIN_DEADLINE,
-                    force_cancel_deadline: Duration::from_secs(2),
-                    begin_shutdown: || Ok(()),
-                    reload_runtime_view: || Ok(()),
-                    finalize_shutdown: || {},
-                    publish_ready: move || {
-                        ready_tx.send(()).map_err(|_| {
-                            AtmError::daemon_unavailable(
-                                "shutdown rejection test failed to observe the daemon ready signal",
-                            )
-                            .with_recovery(
-                                "Restore the bounded ready-signal handshake before retrying the same-host daemon shutdown rejection test.",
-                            )
-                        })
-                    },
-                },
-            );
-            serve_result_tx.send(result).expect("send serve result");
-        });
-
-        lifecycle.terminate_flag().store(true, Ordering::SeqCst);
-        let mut stream = connect_daemon_local_ipc_until_ready(&socket_path, ready_rx);
-        stream
-            .set_send_timeout(Some(Duration::from_secs(5)))
-            .expect("set send timeout");
-        stream
-            .set_recv_timeout(Some(Duration::from_secs(5)))
-            .expect("set recv timeout");
-        let request = RequestEnvelope::Doctor(DoctorQuery {
-            home_dir: tempdir.path().join("home"),
-            current_dir: tempdir.path().join("cwd"),
-            team_override: None,
-        });
-        let request_id = atm_core::protocol::next_request_id();
-        let frame =
-            atm_core::protocol::request_to_frame_payload(request_id, request).expect("frame");
-        atm_core::protocol::write_frame(&mut stream, &frame, "write doctor frame").expect("write");
-        std::io::Write::flush(&mut stream).expect("flush");
-        let response_frame = atm_core::protocol::read_frame(
-            &mut stream,
-            "read shutdown response frame",
-            "shutdown response frame too large",
-        )
-        .expect("read frame")
-        .expect("response frame");
-        let (response_id, response) =
-            atm_core::protocol::response_from_frame_payload(response_frame)
-                .expect("decode response");
-        assert_eq!(response_id, request_id);
-        match response {
-            ResponseEnvelope::Error(error) => {
-                assert_eq!(error.code, AtmErrorCode::DaemonUnavailable);
-                assert!(error.message.contains("shutting down"));
-            }
-            other => panic!("unexpected shutdown response: {other:?}"),
-        }
-
-        serve_result_rx
-            .recv_timeout(Duration::from_secs(5))
-            .expect("recv serve result")
-            .expect("serve runtime result");
-        join.join().expect("join serve thread");
-    }
-
-    #[cfg(unix)]
-    #[test]
-    #[serial_test::serial(env)]
-    fn panic_in_dispatch_still_cleans_up_socket_path_on_shutdown() {
-        let tempdir = TempDir::new().expect("tempdir");
-        let atm_home = tempdir.path().join("atm-home");
-        std::fs::create_dir_all(&atm_home).expect("atm home dir");
-        let _env = EnvGuard::set_many([
-            ("ATM_HOME", Some(atm_home.to_str().expect("utf8 atm home"))),
-            ("ATM_LOG_DIR", None),
-            ("HOME", Some(tempdir.path().to_str().expect("utf8 home"))),
-            ("USERPROFILE", None),
-        ]);
-        let socket_path = tempdir.path().join("daemon.sock");
-        let mut runtime =
-            PreparedRuntimeServer::bind(socket_path.clone()).expect("prepare runtime");
-        let endpoint_guard = runtime.take_endpoint_guard().expect("take endpoint guard");
-        let lifecycle = LifecycleControlSourceAdapter::install().expect("install lifecycle");
-        let _reset = LifecycleFlagResetGuard::install(lifecycle.clone());
-        let dispatcher: Arc<dyn RequestDispatcher + Send + Sync> = Arc::new(PanicDispatcher);
-        let (serve_result_tx, serve_result_rx) = mpsc::channel();
-        let (ready_tx, ready_rx) = mpsc::sync_channel(1);
-
-        let join = std::thread::spawn(move || {
-            let result = runtime.serve_with_runtime_hooks(
-                dispatcher,
-                RuntimeServeHooks {
-                    endpoint_guard,
-                    graceful_drain_deadline: TEST_GRACEFUL_DRAIN_DEADLINE,
-                    force_cancel_deadline: Duration::from_secs(2),
-                    begin_shutdown: || Ok(()),
-                    reload_runtime_view: || Ok(()),
-                    finalize_shutdown: || {},
-                    publish_ready: move || {
-                        ready_tx.send(()).map_err(|_| {
-                            AtmError::daemon_unavailable(
-                                "panic recovery test failed to observe the daemon ready signal",
-                            )
-                            .with_recovery(
-                                "Restore the bounded ready-signal handshake before retrying the same-host panic recovery test.",
-                            )
-                        })
-                    },
-                },
-            );
-            serve_result_tx.send(result).expect("send serve result");
-        });
-
-        let mut stream = connect_daemon_local_ipc_until_ready(&socket_path, ready_rx);
-        stream
-            .set_send_timeout(Some(Duration::from_secs(5)))
-            .expect("set send timeout");
-        stream
-            .set_recv_timeout(Some(Duration::from_secs(5)))
-            .expect("set recv timeout");
-        let request = RequestEnvelope::Doctor(DoctorQuery {
-            home_dir: tempdir.path().join("home"),
-            current_dir: tempdir.path().join("cwd"),
-            team_override: None,
-        });
-        let request_id = atm_core::protocol::next_request_id();
-        let frame =
-            atm_core::protocol::request_to_frame_payload(request_id, request).expect("frame");
-        atm_core::protocol::write_frame(&mut stream, &frame, "write doctor frame").expect("write");
-        std::io::Write::flush(&mut stream).expect("flush");
-        let response_frame = atm_core::protocol::read_frame(
-            &mut stream,
-            "read panic response",
-            "panic frame too large",
-        )
-        .expect("read frame")
-        .expect("response frame");
-        let (response_id, response) =
-            atm_core::protocol::response_from_frame_payload(response_frame)
-                .expect("decode response");
-        assert_eq!(response_id, request_id);
-        assert!(matches!(response, ResponseEnvelope::Error(_)));
-
-        lifecycle.set_terminate_for_test(true);
-        // Two independent call sites race to reap the same panicked dispatch-worker
-        // JoinHandle: the per-connection reap in `handle_connection`, and the opportunistic
-        // accept-loop reap in `prepare_accept_iteration` (which always runs on the very
-        // iteration that observes the termination signal, before the loop breaks). Both are
-        // now non-escalating on a dispatch-worker panic (see
-        // `ActiveConnectionRegistry::reap_finished_dispatches`), so by the time the shutdown
-        // drain's own escalating reap (`join_tracked_dispatches`) runs, the panicked handle
-        // has always already been removed from the tracked list. The serve result is
-        // therefore deterministically non-fatal for this single-request-panic scenario,
-        // regardless of which opportunistic reap wins the race; only a panic first
-        // discovered during the deliberate shutdown drain is escalated as fatal (covered by
-        // `drain_active_connections_for_shutdown_surfaces_dispatch_panic_as_fatal` and
-        // `active_connection_registry::tests::join_tracked_dispatches_escalates_after_panic`).
-        serve_result_rx
-            .recv_timeout(Duration::from_secs(15))
-            .expect("recv serve result")
-            .expect("serve runtime result");
-        join.join().expect("join serve thread");
-        assert!(
-            !socket_path.exists(),
-            "socket endpoint should be removed during shutdown even after a dispatch panic"
         );
     }
